changed Team class attributes to match pep8, added deprecation warnings. TODO: tests for deprecation warnings, tests for espnff.py
