+++ espnff/utils.py
@@ -1,4 +1,5 @@
 import numpy
+import warnings
 
 
 def square_matrix(X):
@@ -45,8 +46,24 @@ def power_points(dominance, teams, week):
         avg_score = sum(team.scores[:week]) / week
         avg_mov = sum(team.mov[:week]) / week
 
-        power = '{0:.2f}'.format((int(i)*0.8) + (int(avg_score)*0.15) + 
+        power = '{0:.2f}'.format((int(i)*0.8) + (int(avg_score)*0.15) +
                                  (int(avg_mov)*0.05))
         power_points.append(power)
-    power_tup = [(i, j) for (i,j) in zip(power_points, teams)]
+    power_tup = [(i, j) for (i, j) in zip(power_points, teams)]
     return sorted(power_tup, key=lambda tup: float(tup[0]), reverse=True)
+
+
+def deprecated_property(old_name, new_name):
+    def getter(self):
+        warnings.warn('%s will be removed in future versions, please use %s instead.'
+                      % (old_name, new_name), stacklevel=2)
+        return getattr(self, new_name)
+
+    def setter(self, value):
+        warnings.warn('%s will be removed in future versions, please use %s instead.'
+                      % (old_name, new_name), stacklevel=2)
+        setattr(self, new_name, value)
+
+    prop = property(getter)
+    prop.setter(setter)
+    return prop
